# ARROW-8761: [C++] Improve the performance of minmax kernel