We improve the performance of the max-min kernel with the simple idea: if the current value is smaller than the current min value; then there is no need to compare it against the current max value, because it must be smaller than the current max value.

This simple trick reduces the expected number of comparisons from 2n to 1.5n, which can be notable for large arrays.