@pitrou 2020-05-11
Can you post any benchmark numbers? Otherwise, we cannot be sure that this will actually improve performance, especially as you're introducing some branches.