@wesm 2020-05-11
I think it will be OK to merge this but _please_ , I beg anyone reading, please limit PRs to src/arrow/compute until I get my (very large) refactor PR up (should be this week 🤞 🤞 )