@wesm 2020-05-11
If you all would mind holding off on merging any patch related to this for a few days or a week I would be appreciative. I can help with rebasing as is needed