@liyafan82 2020-05-12
Hi all,

Thanks a lot for your feedback and good comments.

I have added two benchmarks for the minmax kernel, one for integers and the other for floating point numbers.

The result for integers are as follows:

Before
```
MinMaxIntKernel/32768/0          2.99 us         2.99 us       234012 bytes_per_second=20.387G/s null_percent=0 size=32.768k
MinMaxIntKernel/32768/1          25.7 us         25.7 us        27235 bytes_per_second=2.37282G/s null_percent=1 size=32.768k
MinMaxIntKernel/32768/10         32.4 us         32.5 us        21578 bytes_per_second=1.88089G/s null_percent=10 size=32.768k
MinMaxIntKernel/32768/50         67.7 us         67.6 us        10470 bytes_per_second=924.789M/s null_percent=50 size=32.768k
```

After
```
MinMaxIntKernel/32768/0          7.63 us         7.62 us        92401 bytes_per_second=8.00667G/s null_percent=0 size=32.768k
MinMaxIntKernel/32768/1          25.7 us         25.8 us        27080 bytes_per_second=2.36864G/s null_percent=1 size=32.768k
MinMaxIntKernel/32768/10         34.0 us         34.0 us        20539 bytes_per_second=1.79494G/s null_percent=10 size=32.768k
MinMaxIntKernel/32768/50         68.6 us         68.6 us        10205 bytes_per_second=910.435M/s null_percent=50 size=32.768k
```

It can be seen that when there the null percent is 0, our new method is 2.55x slower, due to the optimizations discussed by @cyb70289. When the null percent is over 0 (even it is 1%), the two methods show almost no performance difference. 

The results for floating point numbers are as follows:

Before
```
MinMaxFloatKernel/32768/0        60.3 us         60.4 us        11765 bytes_per_second=1035.2M/s null_percent=0 size=32.768k
MinMaxFloatKernel/32768/1        96.2 us         96.3 us         7356 bytes_per_second=648.745M/s null_percent=1 size=32.768k
MinMaxFloatKernel/32768/10       95.7 us         95.8 us         7310 bytes_per_second=652.737M/s null_percent=10 size=32.768k
MinMaxFloatKernel/32768/50       97.3 us         97.4 us         7186 bytes_per_second=641.419M/s null_percent=50 size=32.768k
```

After 
```
MinMaxFloatKernel/32768/0        28.8 us         28.9 us        24261 bytes_per_second=2.11477G/s null_percent=0 size=32.768k
MinMaxFloatKernel/32768/1        29.4 us         29.4 us        23759 bytes_per_second=2.07604G/s null_percent=1 size=32.768k
MinMaxFloatKernel/32768/10       38.2 us         38.2 us        18367 bytes_per_second=1.59694G/s null_percent=10 size=32.768k
MinMaxFloatKernel/32768/50       72.4 us         72.5 us         9624 bytes_per_second=862.128M/s null_percent=50 size=32.768k
```

It can be seem that for different null percentage, the new method is consistently faster.