@liyafan82 2020-05-12
> @liyafan82 Did you compile in release mode?

Sure. We should never use debug mode for perfromance evaluations.